WIP character skill progression: - Fixing an item increases the skill required to fix the item. - Fabricating something increases the skills required to fabricate the item. - Repairing a wall increases the mechanic skill.

+++ Barotrauma/BarotraumaShared/Source/Characters/Animation/HumanoidAnimController.cs
@@ -945,7 +945,7 @@ namespace Barotrauma
                 target.Oxygen += deltaTime * 0.5f; //Stabilize them                
             }
 
-            int skill = character.GetSkillLevel("Medical");
+            int skill = (int)character.GetSkillLevel("Medical");
             if (cprAnimState % 17 > 15.0f && targetHead != null && head != null)
             {
                 float yPos = (float)Math.Sin(cprAnimState) * 0.2f;
